requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.0
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.0.1
python-dateutil==2.8.2
reportlab==4.0.9  # For PDF generation
pandas==2.1.3     # For data manipulation
openpyxl==3.1.2   # For Excel export
sqlalchemy==2.0.23
gunicorn==21.2.0  # For production